수정중

+++ src/main/java/com/psm/api/workload/entity/AvailableActionEntity.java
@@ -22,7 +22,7 @@ public class AvailableActionEntity {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private long idx;
+	private int idx;
 	
 	@Column(nullable = false)
 	private String workloadId;
